requirements.txt
opencv-python-headless==4.9.0.80
Pillow==10.2.0
Requests==2.31.0
streamlit==1.29.0
translate==3.6.1

